GPU: Only select drivers which support our shader formats
There's no sense returning a driver if it doesn't support any of the
shader formats we provided. We're at least checking this if a driver is
manually requested, as of #199, but we never seem to do so if we're
otherwise looking for a default driver.

With this, we'll at least, e.g., get a Vulkan backend on Windows if we
only provide SPIR-V shaders, rather than getting an error because the
default D3D12 driver doesn't support our shader format.